Added most of the business logic of the project
Added the data access script, but we found it wasn't working with the
actual database when we integrated it with the rest of the project, but
it worked well independently. Instead we have a small amount of test
code that grabs unrelated images for our representative image
manipulation scripts.

Added building and running dockerfiles, as well as logging image build
output. The pipeline works fairly well, but there is still a problem at
the end, when running the containers. The issue is with the fact that
our manager container is connected to our host machine's docker daemon,
so when paths are specified to create input/output volumes, these are
taken to be paths on the host machine, whereas we had expected them to
mean paths on the manager's container. Barring this issue, the rest of
the project does work successfully; we tested without I/O volumes and
the created containers did perform correctly.